+++ src/main/scala/matt/definitions/Generic.scala
@@ -8,8 +8,8 @@ import matt.POI;
 import matt.SpatialObject;
 
 object Generic {
-  def IsInNodeRegion(long: Float, lat: Float, node: Int, gridSize: Float, cellSize: Float, minLong: Float
-                     , maxLat: Float, width: Int): Boolean = {
+  def IsInNodeRegion(long: Double, lat: Double, node: Int, gridSize: Double, cellSize: Double, minLong: Double
+                     , maxLat: Double, width: Int): Boolean = {
     val nodeJ = ((node - 1) / width).toInt
     val nodeI = node - nodeJ * width - 1
     if (nodeI * gridSize + minLong <= long && long < (nodeI + 1) * gridSize + minLong + cellSize)
@@ -19,7 +19,7 @@ object Generic {
   }
 
 
-  def poiToKeyValue(row: Row, width: Int, minmaxLong: (Float, Float), minmaxLat: (Float, Float), eps: Float
+  def poiToKeyValue(row: Row, width: Int, minmaxLong: (Double, Double), minmaxLat: (Double, Double), eps: Double
                     , geometryFactory: GeometryFactory, gridIndexer: GridIndexer): Seq[(Int, POI)] = {
     var gridSize = 0.0
     if ((minmaxLong._2 - minmaxLong._1) > (minmaxLat._2 - minmaxLat._1))
@@ -29,16 +29,16 @@ object Generic {
     val cellSize = gridSize * eps
 
     val keywords = row.getAs[String]("keywords").split(",").toList;
-    val long = row.getAs[Float]("longtitude")
-    val lat = row.getAs[Float]("latitude")
+    val long = row.getAs[Double]("longtitude")
+    val lat = row.getAs[Double]("latitude")
     //val newNode = extractNode(x.get(0), x.get(1), width, minmaxLong, minmaxLat);
     val result = ListBuffer[(Int, POI)]()
    // result.add((1,new POI(row.getAs[String]("id"), row.getAs[String]("name")
-   //   , row.getAs[Float]("longtitude"), row.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+   //   , row.getAs[Double]("longtitude"), row.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
    // return result
-    return gridIndexer.getNodeIndex(row.getAs[Float]("longtitude"), row.getAs[Float]("latitude"))
+    return gridIndexer.getNodeIndex(row.getAs[Double]("longtitude"), row.getAs[Double]("latitude"))
       .map(x => (x._2 * width + x._1 + 1, new POI(row.getAs[String]("id"), row.getAs[String]("name")
-        , row.getAs[Float]("longtitude"), row.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+        , row.getAs[Double]("longtitude"), row.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
   }
 
   def intersects(point1: SpatialObject, point2: SpatialObject): Boolean = {
@@ -55,25 +55,25 @@ object Generic {
   }
 }
 /*    for(node <-1 until (width*width)+1 ){
-      if(IsInNodeRegion(long,lat,node,gridSize.asInstanceOf[Float],cellSize.asInstanceOf[Float],minmaxLong._1,minmaxLat._2,width)){
+      if(IsInNodeRegion(long,lat,node,gridSize.asInstanceOf[Double],cellSize.asInstanceOf[Double],minmaxLong._1,minmaxLat._2,width)){
 
       }
-     //   result.add((node, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+     //   result.add((node, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
    }
     //
    /* val regionI=((long - minmaxLong._1) / gridSize).toInt
     val regionJ=((lat - minmaxLat._1) / gridSize).toInt
-    result.add((width*regionJ + regionI, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+    result.add((width*regionJ + regionI, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
     if(long-(minmaxLong._1+regionI*gridSize)<cellSize&&(minmaxLat._1+(regionJ+1)*gridSize)-lat<cellSize&&regionI>0&&regionJ<width-1){
-      result.add((width*regionJ + (regionI-1), new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
-      result.add((width*(regionJ+1) + regionI, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
-      result.add((width*(regionJ+1) + (regionI-1), new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+      result.add((width*regionJ + (regionI-1), new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
+      result.add((width*(regionJ+1) + regionI, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
+      result.add((width*(regionJ+1) + (regionI-1), new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
     }
     else if(long-(minmaxLong._1+regionI*gridSize)<cellSize&&regionI>0){
-      result.add((width*regionJ + (regionI-1), new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+      result.add((width*regionJ + (regionI-1), new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
     }
     else if((minmaxLat._1+(regionJ+1)*gridSize)-lat<cellSize&&regionJ<width-1){
-      result.add((width*(regionJ+1) + regionI, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Float]("longtitude"), x.getAs[Float]("latitude"), keywords, 0, geometryFactory)))
+      result.add((width*(regionJ+1) + regionI, new POI(x.getAs[String]("id"), x.getAs[String]("name"), x.getAs[Double]("longtitude"), x.getAs[Double]("latitude"), keywords, 0, geometryFactory)))
     }*/
 
     result*/
